Reorder the arguments of sprite draw
The order of sprite's draw arguments defined what to draw and where.
But the order prevents using sprites and shapes interchangeably with
duck typing. Sprites and shapes have in common the where and the rest
are specific like sheet location, width, height, radius and more.

So the sprite's arguments were reordered to better match the where then
what ordering required by shapes.